third_party/WebKit/LayoutTests/imported/web-platform-tests/webrtc/rtcpeerconnection/rtcpeerconnection-idl-expected.txt

This is a testharness.js-based test.
FAIL RTCPeerConnection interface: existence and properties of interface object assert_false: self's property "RTCPeerConnection" is enumerable expected false got true
PASS RTCPeerConnection interface object length 
FAIL RTCPeerConnection interface: existence and properties of interface prototype object Cannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnection interface: existence and properties of interface prototype object's "constructor" property 
FAIL RTCPeerConnection interface: operation createOffer(RTCOfferOptions) assert_equals: property has wrong .length expected 0 but got 1
FAIL RTCPeerConnection interface: operation createAnswer() assert_equals: property has wrong .length expected 0 but got 1
PASS RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescription) 
PASS RTCPeerConnection interface: attribute localDescription 
PASS RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescription) 
PASS RTCPeerConnection interface: attribute remoteDescription 
PASS RTCPeerConnection interface: attribute signalingState 
FAIL RTCPeerConnection interface: operation updateIce(RTCConfiguration) assert_equals: property has wrong .length expected 1 but got 0
PASS RTCPeerConnection interface: operation addIceCandidate(RTCIceCandidate) 
PASS RTCPeerConnection interface: attribute iceGatheringState 
PASS RTCPeerConnection interface: attribute iceConnectionState 
FAIL RTCPeerConnection interface: attribute canTrickleIceCandidates assert_true: The prototype object must have a property "canTrickleIceCandidates" expected true got false
FAIL RTCPeerConnection interface: operation getConfiguration() assert_own_property: interface prototype object missing non-static operation expected property "getConfiguration" missing
PASS RTCPeerConnection interface: operation close() 
PASS RTCPeerConnection interface: attribute onnegotiationneeded 
PASS RTCPeerConnection interface: attribute onicecandidate 
PASS RTCPeerConnection interface: attribute onsignalingstatechange 
PASS RTCPeerConnection interface: attribute oniceconnectionstatechange 
FAIL RTCPeerConnection interface: attribute onicegatheringstatechange assert_true: The prototype object must have a property "onicegatheringstatechange" expected true got false
FAIL RTCPeerConnection interface: operation createOffer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback,RTCOfferOptions) assert_equals: property has wrong .length expected 0 but got 1
PASS RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescription,VoidFunction,RTCPeerConnectionErrorCallback) 
FAIL RTCPeerConnection interface: operation createAnswer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback) assert_equals: property has wrong .length expected 0 but got 1
PASS RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescription,VoidFunction,RTCPeerConnectionErrorCallback) 
PASS RTCPeerConnection interface: operation addIceCandidate(RTCIceCandidate,VoidFunction,RTCPeerConnectionErrorCallback) 
FAIL RTCPeerConnection interface: operation getSenders() assert_own_property: interface prototype object missing non-static operation expected property "getSenders" missing
FAIL RTCPeerConnection interface: operation getReceivers() assert_own_property: interface prototype object missing non-static operation expected property "getReceivers" missing
FAIL RTCPeerConnection interface: operation addTrack(MediaStreamTrack,MediaStream) assert_own_property: interface prototype object missing non-static operation expected property "addTrack" missing
FAIL RTCPeerConnection interface: operation removeTrack(RTCRtpSender) assert_own_property: interface prototype object missing non-static operation expected property "removeTrack" missing
FAIL RTCPeerConnection interface: attribute ontrack assert_true: The prototype object must have a property "ontrack" expected true got false
PASS RTCPeerConnection interface: operation createDataChannel(DOMString,RTCDataChannelInit) 
PASS RTCPeerConnection interface: attribute ondatachannel 
PASS RTCPeerConnection interface: operation createDTMFSender(MediaStreamTrack) 
FAIL RTCPeerConnection interface: operation getStats(MediaStreamTrack,RTCStatsCallback,RTCPeerConnectionErrorCallback) assert_equals: property has wrong .length expected 3 but got 1
FAIL RTCPeerConnection interface: operation setIdentityProvider(DOMString,DOMString,DOMString) assert_own_property: interface prototype object missing non-static operation expected property "setIdentityProvider" missing
FAIL RTCPeerConnection interface: operation getIdentityAssertion() assert_own_property: interface prototype object missing non-static operation expected property "getIdentityAssertion" missing
FAIL RTCPeerConnection interface: attribute peerIdentity assert_true: The prototype object must have a property "peerIdentity" expected true got false
FAIL RTCPeerConnection interface: attribute onidentityresult assert_true: The prototype object must have a property "onidentityresult" expected true got false
FAIL RTCPeerConnection interface: attribute onpeeridentity assert_true: The prototype object must have a property "onpeeridentity" expected true got false
FAIL RTCPeerConnection interface: attribute onidpassertionerror assert_true: The prototype object must have a property "onidpassertionerror" expected true got false
FAIL RTCPeerConnection interface: attribute onidpvalidationerror assert_true: The prototype object must have a property "onidpvalidationerror" expected true got false
PASS RTCPeerConnection must be primary interface of pc 
FAIL Stringification of pc Cannot read property 'has_stringifier' of undefined
PASS RTCPeerConnection interface: pc must inherit property "createOffer" with the proper type (0) 
PASS RTCPeerConnection interface: calling createOffer(RTCOfferOptions) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "createAnswer" with the proper type (1) 
PASS RTCPeerConnection interface: pc must inherit property "setLocalDescription" with the proper type (2) 
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescription) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "localDescription" with the proper type (3) 
PASS RTCPeerConnection interface: pc must inherit property "setRemoteDescription" with the proper type (4) 
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescription) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "remoteDescription" with the proper type (5) 
FAIL RTCPeerConnection interface: pc must inherit property "signalingState" with the proper type (6) Unrecognized type RTCSignalingState
PASS RTCPeerConnection interface: pc must inherit property "updateIce" with the proper type (7) 
FAIL RTCPeerConnection interface: calling updateIce(RTCConfiguration) on pc with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () { [native code] }" did not throw
PASS RTCPeerConnection interface: pc must inherit property "addIceCandidate" with the proper type (8) 
PASS RTCPeerConnection interface: calling addIceCandidate(RTCIceCandidate) on pc with too few arguments must throw TypeError 
FAIL RTCPeerConnection interface: pc must inherit property "iceGatheringState" with the proper type (9) Unrecognized type RTCIceGatheringState
FAIL RTCPeerConnection interface: pc must inherit property "iceConnectionState" with the proper type (10) Unrecognized type RTCIceConnectionState
FAIL RTCPeerConnection interface: pc must inherit property "canTrickleIceCandidates" with the proper type (11) assert_inherits: property "canTrickleIceCandidates" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "getConfiguration" with the proper type (12) assert_inherits: property "getConfiguration" not found in prototype chain
PASS RTCPeerConnection interface: pc must inherit property "close" with the proper type (13) 
FAIL RTCPeerConnection interface: pc must inherit property "onnegotiationneeded" with the proper type (14) Unrecognized type EventHandler
FAIL RTCPeerConnection interface: pc must inherit property "onicecandidate" with the proper type (15) Unrecognized type EventHandler
FAIL RTCPeerConnection interface: pc must inherit property "onsignalingstatechange" with the proper type (16) Unrecognized type EventHandler
FAIL RTCPeerConnection interface: pc must inherit property "oniceconnectionstatechange" with the proper type (17) Unrecognized type EventHandler
FAIL RTCPeerConnection interface: pc must inherit property "onicegatheringstatechange" with the proper type (18) assert_inherits: property "onicegatheringstatechange" not found in prototype chain
PASS RTCPeerConnection interface: pc must inherit property "createOffer" with the proper type (19) 
PASS RTCPeerConnection interface: calling createOffer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback,RTCOfferOptions) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "setLocalDescription" with the proper type (20) 
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescription,VoidFunction,RTCPeerConnectionErrorCallback) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "createAnswer" with the proper type (21) 
PASS RTCPeerConnection interface: calling createAnswer(RTCSessionDescriptionCallback,RTCPeerConnectionErrorCallback) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "setRemoteDescription" with the proper type (22) 
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescription,VoidFunction,RTCPeerConnectionErrorCallback) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "addIceCandidate" with the proper type (23) 
PASS RTCPeerConnection interface: calling addIceCandidate(RTCIceCandidate,VoidFunction,RTCPeerConnectionErrorCallback) on pc with too few arguments must throw TypeError 
FAIL RTCPeerConnection interface: pc must inherit property "getSenders" with the proper type (24) assert_inherits: property "getSenders" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "getReceivers" with the proper type (25) assert_inherits: property "getReceivers" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "addTrack" with the proper type (26) assert_inherits: property "addTrack" not found in prototype chain
FAIL RTCPeerConnection interface: calling addTrack(MediaStreamTrack,MediaStream) on pc with too few arguments must throw TypeError assert_inherits: property "addTrack" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "removeTrack" with the proper type (27) assert_inherits: property "removeTrack" not found in prototype chain
FAIL RTCPeerConnection interface: calling removeTrack(RTCRtpSender) on pc with too few arguments must throw TypeError assert_inherits: property "removeTrack" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "ontrack" with the proper type (28) assert_inherits: property "ontrack" not found in prototype chain
PASS RTCPeerConnection interface: pc must inherit property "createDataChannel" with the proper type (29) 
PASS RTCPeerConnection interface: calling createDataChannel(DOMString,RTCDataChannelInit) on pc with too few arguments must throw TypeError 
FAIL RTCPeerConnection interface: pc must inherit property "ondatachannel" with the proper type (30) Unrecognized type EventHandler
PASS RTCPeerConnection interface: pc must inherit property "createDTMFSender" with the proper type (31) 
PASS RTCPeerConnection interface: calling createDTMFSender(MediaStreamTrack) on pc with too few arguments must throw TypeError 
PASS RTCPeerConnection interface: pc must inherit property "getStats" with the proper type (32) 
PASS RTCPeerConnection interface: calling getStats(MediaStreamTrack,RTCStatsCallback,RTCPeerConnectionErrorCallback) on pc with too few arguments must throw TypeError 
FAIL RTCPeerConnection interface: pc must inherit property "setIdentityProvider" with the proper type (33) assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: calling setIdentityProvider(DOMString,DOMString,DOMString) on pc with too few arguments must throw TypeError assert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "getIdentityAssertion" with the proper type (34) assert_inherits: property "getIdentityAssertion" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "peerIdentity" with the proper type (35) assert_inherits: property "peerIdentity" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "onidentityresult" with the proper type (36) assert_inherits: property "onidentityresult" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "onpeeridentity" with the proper type (37) assert_inherits: property "onpeeridentity" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "onidpassertionerror" with the proper type (38) assert_inherits: property "onidpassertionerror" not found in prototype chain
FAIL RTCPeerConnection interface: pc must inherit property "onidpvalidationerror" with the proper type (39) assert_inherits: property "onidpvalidationerror" not found in prototype chain
Harness: the test ran to completion.
